Move DRC-specific GC compiler bits to its own submodule
This lets us differentiate between common code that is needed for any GC
compiler implementation and DRC-specific code.
